+++ pybillomat/invoices.py
@@ -407,7 +407,7 @@ class Invoices(list):
         # Parse XML
         invoices_etree = ET.fromstring(response.data)
 
-        self.per_page = int(invoices_etree.attrib.get("per_page", "0"))
+        self.per_page = int(invoices_etree.attrib.get("per_page", "100"))
         self.total = int(invoices_etree.attrib.get("total", "0"))
         self.page = int(invoices_etree.attrib.get("page", "1"))
         self.pages = (self.total // self.per_page) + int(bool(self.total % self.per_page))